@@ -3,20 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
-    "@material-ui/styles": "^4.11.5",
+    "@date-io/dayjs": "^3.2.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^7.0.1",
+    "@mui/lab": "^7.0.0-beta.10",
+    "@mui/material": "^7.0.1",
+    "@mui/x-date-pickers": "^7.28.3",
+    "@mui/x-date-pickers-pro": "^7.28.3",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.8.4",
-    "connected-react-router": "^6.9.3",
+    "chart.js": "^4.4.8",
+    "dayjs": "^1.11.13",
     "history": "^4.7.2",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3",
-    "react-redux": "^7.1.0",
-    "react-router": "^5.0.0",
+    "react": "18.3.1",
+    "react-chartjs-2": "^5.3.0",
+    "react-dom": "18.3.1",
+    "react-redux": "^8.1.3",
+    "react-router-dom": "^6.30.0",
     "react-scripts": "^5.0.1",
     "redux": "^4.0.0",
     "redux-actions": "^3.0.3",
@@ -53,6 +59,7 @@
     "@types/react": "17.0.0"
   },
   "devDependencies": {
-    "@redux-devtools/extension": "^3.3.0"
+    "@redux-devtools/extension": "^3.3.0",
+    "@testing-library/react": "^14.3.1"
   }
 }
