@@ -1,36 +1,36 @@
-import { createTheme } from '@material-ui/core/styles';
+import { createTheme } from '@mui/material/styles'; // 変更: v5 の createTheme をインポート
 
 // Pick colors on https://material.io/resources/color/#!/
 
 export const theme = createTheme({
-  palette: {
-    primary: {
-      light: '#88ffff',
-      main: '#4dd0e1',
-      dark: '#009faf',
-      contrastText: '#000',
+    palette: {
+        primary: {
+            light: '#88ffff',
+            main: '#4dd0e1',
+            dark: '#009faf',
+            contrastText: '#000',
+        },
+        secondary: {
+            light: '#ffff81',
+            main: '#ffd54f',
+            dark: '#c8a415',
+            contrastText: '#000',
+        },
+        grey: {
+            50: "#fafafa",
+            100: "#f5f5f5",
+            200: "#eeeeee",
+            300: "#e0e0e0",
+            400: "#bdbdbd",
+            500: "#9e9e9e",
+            600: "#757575",
+            700: "#616161",
+            800: "#424242",
+            900: "#212121",
+            A100: "#d5d5d5",
+            A200: "#aaaaaa",
+            A400: "#303030",
+            A700: "#616161"
+        }
     },
-    secondary: {
-      light: '#ffff81',
-      main: '#ffd54f',
-      dark: '#c8a415',
-      contrastText: '#000',
-    },
-    grey: {
-      50: "#fafafa",
-      100: "#f5f5f5",
-      200: "#eeeeee",
-      300: "#e0e0e0",
-      400: "#bdbdbd",
-      500: "#9e9e9e",
-      600: "#757575",
-      700: "#616161",
-      800: "#424242",
-      900: "#212121",
-      A100: "#d5d5d5",
-      A200: "#aaaaaa",
-      A400: "#303030",
-      A700: "#616161"
-    }
-  },
 });
