@@ -1,14 +1,13 @@
 import React, { useState, useCallback } from "react";
-import IconButton from "@material-ui/core/IconButton";
-import MenuItem from "@material-ui/core/MenuItem";
-import Menu from "@material-ui/core/Menu";
+import { IconButton, MenuItem, Menu, Avatar } from "@mui/material";
 import { useDispatch } from "react-redux";
 import { signOut } from "../../reducks/users/operations";
-import { push } from "connected-react-router";
 import guestuserimage from "../../assets/img/guestuserimage.png";
+import { useNavigate } from 'react-router-dom';
 
 const HeaderMenu = (props) => {
     const dispatch = useDispatch();
+    const navigate = useNavigate();
 
     const [anchorEl, setAnchorEl] = useState(null);
     const handleMenuToggle = useCallback((event) => {
@@ -21,29 +20,40 @@ const HeaderMenu = (props) => {
 
     const hasUserimage = useCallback(() => {
         return props.username === "ゲストユーザ" ||
-               (props.username !== "ゲストユーザ" &&
+            (props.username !== "ゲストユーザ" &&
                 props.userimage === "");
     }, [props.username, props.userimage]);
+
+    const handleUserDetailClick = () => {
+        navigate('/users/detail');
+        handleClose();
+    };
+
+    const handleSignOutClick = () => {
+        dispatch(signOut());
+        navigate('/');
+        handleClose();
+    };
 
     return (
         <div>
             <IconButton onClick={handleMenuToggle}>
-                {hasUserimage() ? (
-                    <img src={guestuserimage} alt="仮画像" width="50px" height="40px"/>
-                ) : (
-                    <img src={props.userimage} alt="ユーザ" width="50px" height="40px"/>
-                )}
+                <Avatar
+                    alt={props.username}
+                    src={hasUserimage() ? guestuserimage : props.userimage}
+                    sx={{ width: 40, height: 40 }}
+                />
             </IconButton>
             <Menu
                 anchorEl={anchorEl}
                 open={Boolean(anchorEl)}
                 onClose={handleClose}
             >
-                <MenuItem button onClick={() => { dispatch(push('/users/detail')); handleClose(); }}>ユーザ設定</MenuItem>
-                <MenuItem button onClick={() => { dispatch(signOut()); handleClose(); }}>サインアウト</MenuItem>
+                <MenuItem onClick={handleUserDetailClick}>ユーザ設定</MenuItem>
+                <MenuItem onClick={handleSignOutClick}>サインアウト</MenuItem>
             </Menu>
         </div>
-    )
-}
+    );
+};
 
 export default HeaderMenu;
