@@ -1,25 +1,23 @@
-import React from "react";
-import { Button } from "@material-ui/core";
-import { makeStyles } from "@material-ui/styles";
+import React from 'react';
+import Button from '@mui/material/Button';
+import { styled } from '@mui/material/styles';
 
-const useStyles = makeStyles({
-    "button": {
-        fontSize: 16,
-        height: 48,
-        marginButton: 16,
-        width: 256
-    }
-})
+const StyledButton = styled(Button)(({ theme }) => ({
+    margin: theme.spacing(1),
+}));
 
 const PrimaryButton = (props) => {
-    const classes = useStyles()
-    
+    const { label, onClick, color, disabled } = props;
     return (
-        <Button className={classes.button} variant="contained" onClick={() => props.onClick()}>
-            {props.label}
-        </Button>
-    )
-
-}
+        <StyledButton
+            variant="contained"
+            color={color ? color : "primary"}
+            onClick={onClick}
+            disabled={disabled}
+        >
+            {label}
+        </StyledButton>
+    );
+};
 
 export default PrimaryButton;
