@@ -2,29 +2,23 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import createStore from './reducks/store/store';
-import { ConnectedRouter } from 'connected-react-router';
-import { MuiThemeProvider } from '@material-ui/core'
+import { BrowserRouter } from 'react-router-dom';
+import { ThemeProvider } from '@mui/material/styles';
 import { theme } from './assets/theme';
-import * as History from 'history';
 import * as serviceWorker from './serviceWorker';
-
 import App from './App';
 
-const history = History.createBrowserHistory();
-export const store = createStore(history);
+export const store = createStore();
 
 ReactDOM.render(
   <Provider store={store} >
-    <ConnectedRouter history={history} >
-      <MuiThemeProvider theme={theme}>
+    <BrowserRouter>
+      <ThemeProvider theme={theme}>
         <App />
-      </MuiThemeProvider>
-    </ConnectedRouter>
+      </ThemeProvider>
+    </BrowserRouter>
   </Provider>,
   document.getElementById('root'),
 )
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 serviceWorker.unregister();
