@@ -1,31 +1,26 @@
-import React, { useCallback } from "react"
+import React, { useCallback } from "react";
 import { PrimaryButton } from "../components/UIkit";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteUser } from "../reducks/users/operations";
-import { push } from "connected-react-router";
 import { getEmail, getUserId, getUserImage, getUserName } from "../reducks/users/selectors";
-import { makeStyles } from "@material-ui/core";
+import { styled } from "@mui/material/styles";
 import guestuserimage from "../assets/img/guestuserimage.png";
+import Container from '@mui/material/Container';
+import Typography from '@mui/material/Typography';
+import Box from '@mui/material/Box';
+import Avatar from '@mui/material/Avatar';
+import Grid from '@mui/material/Grid';
+import Paper from '@mui/material/Paper';
+import { useNavigate } from 'react-router-dom';
 
-const useStyles = makeStyles((theme) => ({
-    detail: {
-        textAlign: 'left',
-        [theme.breakpoints.down('sm')] : {
-            margin: '0 auto 16px auto',
-            height: 'auto',
-            width: 320
-        },
-        [theme.breakpoints.up('sm')]: {
-            margin: 'o auto',
-            height: 'auto',
-            width: 400
-        }
-    }
+const StyledPaper = styled(Paper)(({ theme }) => ({
+    padding: theme.spacing(2),
+    textAlign: 'center',
 }));
 
 const UserDetail = () => {
-    const classes = useStyles();
     const dispatch = useDispatch();
+    const navigate = useNavigate();
     const selector = useSelector((state) => state);
     const username = getUserName(selector),
           userimage = getUserImage(selector),
@@ -34,41 +29,56 @@ const UserDetail = () => {
 
     const hasUserimage = useCallback(() => {
         return username === "ゲストユーザ" ||
-        　　 (username !== "ゲストユーザ" &&
-            userimage === "");
+            (username !== "ゲストユーザ" &&
+                userimage === "");
     }, [username, userimage]);
 
-    return (
-        <div className="c-section-wrapin">
-            {userid && (
-                <div className="p-grid__row">
-                    {hasUserimage() ? (
-                        <img src={guestuserimage} alt="ゲスト画像" />
-                    ) : (
-                        <img src={userimage} alt="ユーザ画像" />
-                    )}
-                    <div className={classes.detail}>
-                        <div className="module-spacer--small"/>
-                        <p>ユーザ名：{username}</p>
-                        <div className="module-spacer--small"/>
-                        <p>登録されたメールアドレス：{email}</p>
-                        <div className="module-spacer--medium"/>
-                        <div className="center">
-                          <PrimaryButton
-                              label={"ユーザ情報を編集"}
-                              onClick={() => dispatch(push('/users/edit'))}
-                          />
-                          <div className="module-spacer--medium"/>
-                          <PrimaryButton
-                              label={"退会する"}
-                              onClick={() => dispatch(deleteUser())}
-                          />
-                        </div>
-                    </div>
-                </div>
-            )}
-        </div>
-    )
-}
+    const handleDeleteUser = () => {
+        dispatch(deleteUser());
+        navigate('/');
+    };
 
-export default UserDetail
+    return (
+        <Container maxWidth="sm">
+            <Box sx={{ mt: 8, display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
+                <Typography component="h1" variant="h5" gutterBottom>
+                    ユーザ設定
+                </Typography>
+                {userid && (
+                    <StyledPaper sx={{ width: '100%', mt: 2 }}>
+                        <Grid container spacing={2} direction="column" alignItems="center">
+                            <Grid item>
+                                <Avatar
+                                    alt={username}
+                                    src={hasUserimage() ? guestuserimage : userimage}
+                                    sx={{ width: 80, height: 80 }}
+                                />
+                            </Grid>
+                            <Grid item>
+                                <Typography variant="subtitle1">ユーザ名：{username}</Typography>
+                            </Grid>
+                            <Grid item>
+                                <Typography variant="subtitle2" color="textSecondary">登録されたメールアドレス：{email}</Typography>
+                            </Grid>
+                            <Grid item sx={{ mt: 2 }}>
+                                <PrimaryButton
+                                    label={"ユーザ情報を編集"}
+                                    onClick={() => navigate('/users/edit')}
+                                />
+                            </Grid>
+                            <Grid item>
+                                <PrimaryButton
+                                    label={"退会する"}
+                                    onClick={handleDeleteUser}
+                                    color="error"
+                                />
+                            </Grid>
+                        </Grid>
+                    </StyledPaper>
+                )}
+            </Box>
+        </Container>
+    );
+};
+
+export default UserDetail;
