@@ -5,8 +5,10 @@ class Api::V1::MedicinesController < Api::V1::AliasController
   # GET /medicines_with_category
   # Api::V1::MedicinesController#index
   def index
+    # ユーザが登録した薬を全て取得し、関連するカテゴリーの情報も含めてインスタンス変数に格納
     @medicines = current_user.medicines.all.includes(:category)
 
+    # インスタンス変数に格納された薬の情報を単位とカテゴリー名の情報を付加してJSON形式で返す
     medicines_with_category = @medicines.map do |medicine|
       category_name = medicine.category&.name || "カテゴリー未設定"
       unit_value = medicine.unit
@@ -16,23 +18,17 @@ class Api::V1::MedicinesController < Api::V1::AliasController
     render json: medicines_with_category
   end
 
-  #def index
-  #  @medicines = current_user.medicines.all
-  #  render json: @medicines
-  #end
-
-  # GET /medicines/1
-  #def show
-  #  render json: @medicine.as_json(include: :category)
-  #end
-
   # POST /medicines
   def create
     category_name = params[:medicine][:category_name]
     category = current_user.categories.find_by(name: category_name)
 
+    #　カテゴリー名とカテゴリーのIDが一致する場合
     if category
+      #medicine_paramsからcategory_nameを除外する
       medicine_params_without_category_name = medicine_params.except(:category_name)
+
+      #カテゴリーのIDも含めて薬のインスタンスを作成
       @medicine = current_user.medicines.new(medicine_params_without_category_name.merge(category_id: category.id))
 
       if @medicine.save
@@ -50,8 +46,13 @@ class Api::V1::MedicinesController < Api::V1::AliasController
     category_name = params[:medicine][:category_name]
     category = current_user.categories.find_by(name: category_name)
 
+    #　カテゴリー名とカテゴリーのIDが一致する場合
     if category
+
+      #medicine_paramsからcategory_nameを除外する
       medicine_params_without_category_name = medicine_params.except(:category_name)
+
+      #カテゴリーのIDも含めて薬を更新
       if @medicine.update(medicine_params_without_category_name.merge(category_id: category.id))
         render json: @medicine
       else
