@@ -1,7 +1,7 @@
 class CreatePosts < ActiveRecord::Migration[7.2]
   def change
     create_table :posts do |t|
-      t.string :ingestion_amount
+      t.integer :ingestion_amount
       t.string :comment
       t.references :user, foreign_key: true
       t.references :medicine, foreign_key: true
