@@ -1,2 +1,2 @@
 export {default as MedicineImageArea} from './MedicineImageArea'
-export {default as MedicineSummary} from './MedicineSummary'
+export {default as Medicine} from './Medicine'
