@@ -1,7 +1,5 @@
-export { default as TextInput } from './TextInput';
 export { default as PrimaryButton } from './PrimaryButton';
-export { default as UserImageArea } from './UserImageArea';
-export { default as SelectBox } from './SelectBox';
-export {default as ScrollToTop} from './ScrollToTop';
-export {default as Footer} from './Footer';
-export {default as BackGroundImage} from './BackGroundImage';
+export { default as TextInput } from './TextInput';
+export { default as NavigateTabs } from './NavigateTabs';
+export { default as ScrollToTop } from './ScrollToTop';
+
