@@ -3,7 +3,6 @@ import {
     combineReducers,
     applyMiddleware
 } from 'redux';
-import { connectRouter, routerMiddleware } from 'connected-react-router';
 import { composeWithDevTools } from '@redux-devtools/extension';
 import thunk from 'redux-thunk';
 import { UsersReducer } from "../users/reducers";
@@ -11,19 +10,17 @@ import { MedicinesReducer } from "../medicines/reducers";
 import { PostsReducer } from "../posts/reducers";
 import { CategoriesReducer } from '../categories/reducers';
 
-export default function createStore(history) {
+export default function createStore() {
     const rootReducer = combineReducers({
-        router: connectRouter(history),
         users: UsersReducer,
         medicines: MedicinesReducer,
         posts: PostsReducer,
-        categories: CategoriesReducer
+        categories: CategoriesReducer,
     });
 
     const enhancers = composeWithDevTools(
         applyMiddleware(
-            routerMiddleware(history),
-            thunk
+            thunk,
         )
     );
 
