@@ -1,49 +1,63 @@
 import React, { useCallback, useEffect, useState } from "react";
-import { PrimaryButton, SelectBox, TextInput } from "../components/UIkit";
+import { PrimaryButton, TextInput } from "../components/UIkit";
 import { useDispatch, useSelector } from "react-redux";
 import { updateMedicine, fetchMedicines } from "../reducks/medicines/operations"
 import { MedicineImageArea } from "../components/Medicine";
 import { getMedicines } from "../reducks/medicines/selectors";
 import { getCategories } from "../reducks/categories/selectors";
 import { fetchCategories } from "../reducks/categories/operations";
-import { push } from "connected-react-router";
+import { useNavigate, useParams } from 'react-router-dom';
+import Container from '@mui/material/Container';
+import Typography from '@mui/material/Typography';
+import Box from '@mui/material/Box';
+import FormControl from '@mui/material/FormControl';
+import InputLabel from '@mui/material/InputLabel';
+import Select from '@mui/material/Select';
+import MenuItem from '@mui/material/MenuItem';
 
 const MedicineEdit = () => {
     const dispatch = useDispatch();
-
+    const navigate = useNavigate();
+    const { id } = useParams();
     const selector = useSelector((state) => state)
-    const path = selector.router.location.pathname
-    const id = path.split('/medicines/edit/')[1]
-
     const medicines = getMedicines(selector);
     const categories = getCategories(selector);
 
-    const [medicineName, setMedicineName] = useState(""),
-          [medicineImage, setMedicineImage] = useState(""),
-          [memo, setMemo] = useState(""),
-          [unit, setUnit] = useState(""),
-          [ingestionTimesPerDay, setIngestionTimesPerDay] = useState(""),
-          [ingestionAmountEveryTime, setIngestionAmountEveryTime] = useState(""),
-          [categoryName, setCategoryName] = useState("");
+    const [medicineName, setMedicineName] = useState("");
+    const [medicineImage, setMedicineImage] = useState("");
+    const [memo, setMemo] = useState("");
+    const [unit, setUnit] = useState("");
+    const [ingestionTimesPerDay, setIngestionTimesPerDay] = useState("");
+    const [ingestionAmountEveryTime, setIngestionAmountEveryTime] = useState("");
+    const [categoryName, setCategoryName] = useState("");
 
     const inputMedicineName = useCallback((event) => {
-        setMedicineName(event.target.value)
-    }, [setMedicineName])
+        setMedicineName(event.target.value);
+    }, [setMedicineName]);
 
     const inputMemo = useCallback((event) => {
-        setMemo(event.target.value)
-    }, [setMemo])
+        setMemo(event.target.value);
+    }, [setMemo]);
 
     const inputIngestionTimesPerDay = useCallback((event) => {
-        setIngestionTimesPerDay(event.target.value)
-    }, [setIngestionTimesPerDay])
+        setIngestionTimesPerDay(event.target.value);
+    }, [setIngestionTimesPerDay]);
 
     const inputIngestionAmountPerDay = useCallback((event) => {
-        setIngestionAmountEveryTime(event.target.value)
-    }, [setIngestionAmountEveryTime])
+        setIngestionAmountEveryTime(event.target.value);
+    }, [setIngestionAmountEveryTime]);
+
+    const handleCategoryChange = (event) => {
+        setCategoryName(event.target.value);
+    };
+
+    const handleUnitChange = (event) => {
+        setUnit(event.target.value);
+    };
 
     useEffect(() => {
-        dispatch(fetchMedicines())
+        dispatch(fetchMedicines());
+        dispatch(fetchCategories());
     }, [dispatch]);
 
     useEffect(() => {
@@ -55,72 +69,91 @@ const MedicineEdit = () => {
             setIngestionTimesPerDay(String(targetMedicine.ingestion_times_per_day));
             setIngestionAmountEveryTime(String(targetMedicine.ingestion_amount_every_time));
             setCategoryName(targetMedicine.category_name);
+            setMedicineImage(targetMedicine.medicine_image || "");
         }
     }, [id, medicines, setMedicineName, setMemo, setIngestionTimesPerDay, setIngestionAmountEveryTime, setCategoryName]);
 
-    useEffect(() => {
-            dispatch(fetchCategories()); // enum の値をフェッチ
-    }, [dispatch]);
-
-    const categoryOptions = categories ? categories.map((category) => ({
-        label: category.name,
-        value: category.name,
-        key: category.id
-    })) : [];
+    const categoryOptions = categories ? categories.map((category) => (
+        <MenuItem key={category.id} value={category.name}>
+            {category.name}
+        </MenuItem>
+    )) : [];
 
     const unitOptions = [
-        { label: 'g', value: 'g', key: 'g' },
-        { label: 'mg', value: 'mg', key: 'mg' },
-        { label: '錠', value: '錠', key: '錠' },
-        { label: '枚', value: '枚', key: '枚' },
-        { label: 'ml', value: 'ml', key: 'ml' }
+        <MenuItem key="g" value="g">g</MenuItem>,
+        <MenuItem key="mg" value="mg">mg</MenuItem>,
+        <MenuItem key="錠" value="錠">錠</MenuItem>,
+        <MenuItem key="枚" value="枚">枚</MenuItem>,
+        <MenuItem key="ml" value="ml">ml</MenuItem>
     ];
 
-    return(
-        <section>
-            <h2 className="u-text_headline u-text-center">薬の編集</h2>
-            <div className="c-section-container">
-                <MedicineImageArea medicineImage={medicineImage} setMedicineImage={setMedicineImage} />
+    const handleUpdateMedicine = () => {
+        dispatch(updateMedicine(id, medicineName, medicineImage, memo, unit, ingestionTimesPerDay, ingestionAmountEveryTime, categoryName));
+        navigate(`/medicines/${id}`);
+    };
+
+    return (
+        <Container maxWidth="sm">
+            <Box sx={{ mt: 4 }}>
+                <Typography variant="h4" gutterBottom>薬の編集</Typography>
+                <MedicineImageArea medicineImage={medicineImage} setMedicineImage={setMedicineImage}/>
                 <TextInput
-                    fullWidth={true} label={"薬名"} multiline={false} required={true}
+                    fullWidth label={"薬名"} multiline={false} required={true}
                     onChange={inputMedicineName} rows={1} value={medicineName} type={"text"}
+                    sx={{ mt: 2 }}
                 />
-
                 <TextInput
-                    fullWidth={true} label={"メモ"} multiline={true} required={false}
+                    fullWidth label={"メモ"} multiline={true} required={false}
                     onChange={inputMemo} rows={5} value={memo} type={"text"}
+                    sx={{ mt: 2 }}
                 />
-
-                <SelectBox
-                    label={"カテゴリー"} required={true} options={categoryOptions} select={setCategoryName} value={categoryName}
-                />
-
+                <FormControl fullWidth required sx={{ mt: 2 }}>
+                    <InputLabel id="category-label">カテゴリー</InputLabel>
+                    <Select
+                        labelId="category-label"
+                        id="category"
+                        value={categoryName}
+                        label="カテゴリー"
+                        onChange={handleCategoryChange}
+                    >
+                        {categoryOptions}
+                    </Select>
+                </FormControl>
                 <TextInput
-                    fullWidth={true} label={"1日の服薬回数"} multiline={false} required={true}
+                    fullWidth label={"1日の服薬回数"} multiline={false} required={true}
                     onChange={inputIngestionTimesPerDay} rows={1} value={ingestionTimesPerDay} type={"number"}
+                    sx={{ mt: 2 }}
                 />
-
                 <TextInput
-                    fullWidth={true} label={"1回の服薬量"} multiline={false} required={true}
+                    fullWidth label={"1回の服薬量"} multiline={false} required={true}
                     onChange={inputIngestionAmountPerDay} rows={1} value={ingestionAmountEveryTime} type={"number"}
+                    sx={{ mt: 2 }}
                 />
-
-                <SelectBox
-                    label={"単位"} required={true} options={unitOptions} select={setUnit} value={unit}
-                />
-
-                <div className="module-spacer--small"/>
-                <div className="center">
+                <FormControl fullWidth required sx={{ mt: 2 }}>
+                    <InputLabel id="unit-label">単位</InputLabel>
+                    <Select
+                        labelId="unit-label"
+                        id="unit"
+                        value={unit}
+                        label="単位"
+                        onChange={handleUnitChange}
+                    >
+                        {unitOptions}
+                    </Select>
+                </FormControl>
+                <Box sx={{ mt: 3, display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
                     <PrimaryButton
                         label={"薬を更新"}
-                        onClick={() => dispatch(updateMedicine(id, medicineName, medicineImage, memo, unit, ingestionTimesPerDay, ingestionAmountEveryTime, categoryName))}
+                        onClick={handleUpdateMedicine}
+                        sx={{ mb: 1 }}
                     />
-                    <div className="module-spacer--medium"/>
-                    <p onClick={() => dispatch(push('/medicines/'+ id))}>戻る</p>
-                </div>
-            </div>
-        </section>
-    )
-}
+                    <Typography sx={{ mt: 1, cursor: 'pointer' }} onClick={() => navigate(`/medicines/${id}`)}>
+                        戻る
+                    </Typography>
+                </Box>
+            </Box>
+        </Container>
+    );
+};
 
-export default MedicineEdit
+export default MedicineEdit;
